Fix every clang-tidy error rather than muting the checks.
71 of them, and the four families they come from are the ones that stop a
build: correctness, UB, performance, concurrency.

Fifty-two were narrowing conversions, and almost all of one shape -- a
stratified sample position, `((i * 3) % 2048 + 0.5f) / 2048.0f`, written out
forty times across four suites, plus twenty copies of `p[0] / 255.0f`
decoding a UNORM8 byte. Those become tests/support/sampling.h: stratum() and
unorm8(), constexpr, with the conversion explicit inside them. Forty
static_casts would have silenced the same diagnostics and said nothing; a
name says what the number is, and leaves the analyzer able to complain about
the narrowing conversions that are not deliberate.

Two are worth reading rather than skimming:

standard_pbr.h walked its lobe CDF with `u_lobe < (cdf += p_specular)` inside
the else-if chain. Now the four running sums are named before it. Same
additions in the same order, so bit-identical, but the short-circuit is no
longer load-bearing in a way that reads like a typo -- on `dt_only_exit ||`
the increment never ran, which is correct precisely because that branch is
the one being taken.

sharc_grid.h encoded with `(uint32_t)(v * kScale + 0.5f)`. roundf is not a
workaround for the check here: v * kScale + 0.5f can round up to the next
representable float by itself, and the truncation then lands a quantum high.
kClamp keeps today's values clear of it; the new spelling stays correct if
kScale ever grows. The Metal port of this file still has the old form and
still needs the same change, as its own comment already says.

The rest: PushID(nodeId) instead of laundering an index through a void* for
the outliner's widget identity, an integer induction variable for a gizmo
probe that was stepping a float by 5.0f, size_t multiplications where the
result indexes, constexpr on two glm globals, a corrected argument comment.

Five sites keep a NOLINT, because setenv/getenv have no thread-safe standard
alternative: the CLI writes MTL_CAPTURE_ENABLED on the first statement of
main() before any thread exists, and the rest is doctest, which runs one
case at a time. Each says so, matching the convention already in
foundation/env/env.h.

Full run is now 0 errors, 845 warnings, none of them in a family that fails
a build. The OptiX and Metal translation units are in neither
compile_commands.json nor that number, so sharc_grid.h's change is compiled
here only by the host test build.

+++ src/shaders/optix/shading/texture_transform.h
@@ -30,8 +30,7 @@ DEVICE_FUNC float2 apply_texture_transform(float2 uv, const THREAD_REF MaterialP
     // this is what keeps a material that predates that field readable.
     const float kx = (m.uv_scale_x != 0.0f) ? m.uv_scale_x : 1.0f;
     const float ky = (m.uv_scale_y != 0.0f) ? m.uv_scale_y : 1.0f;
-    return make_float2(uv.x * kx * c - uv.y * ky * s + m.uv_offset_x,
-                       uv.x * kx * s + uv.y * ky * c + m.uv_offset_y);
+    return make_float2(uv.x * kx * c - uv.y * ky * s + m.uv_offset_x, uv.x * kx * s + uv.y * ky * c + m.uv_offset_y);
 }
 
 // glTF COLOR_0, packed RGBA8 and LINEAR -- it carries no transfer function,
@@ -39,8 +38,8 @@ DEVICE_FUNC float2 apply_texture_transform(float2 uv, const THREAD_REF MaterialP
 DEVICE_FUNC float3 unpack_vertex_color(unsigned int val)
 {
     const float s = 1.0f / 255.0f;
-    return make_float3((val & 0xffu) * s, ((val >> 8) & 0xffu) * s,
-                       ((val >> 16) & 0xffu) * s);
+    return make_float3(static_cast<float>(val & 0xffu) * s, static_cast<float>((val >> 8) & 0xffu) * s,
+                       static_cast<float>((val >> 16) & 0xffu) * s);
 }
 
 // Coverage of a surface, given the alpha the base-colour slot resolved to.
